pjt/05-pjt/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==4.2.20
django-environ==0.12.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.72.0
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0